Refactor catalog restrictions to own service
Clearer separation of concerns, which also makes it easier to test
